requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
frozenlist==1.6.0
idna==3.10
inflate64==1.0.1
magic-filter==1.0.12
multidict==6.4.3
multivolumefile==0.2.3
patool==4.0.0
propcache==0.3.1
psutil==7.0.0
py7zr==0.22.0
pyaes==1.6.1
pyasn1==0.6.1
pybcj==1.0.3
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyppmd==1.1.1
PySocks==1.7.1
pyzstd==0.16.2
rsa==4.9.1
Telethon==1.40.0
texttable==1.7.0
typing-inspection==0.4.0
typing_extensions==4.13.2
unrar==0.4
yarl==1.20.0
